requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
greenlet==1.1.1
idna==3.2
Markdown==3.3.4
playwright==1.14.1
pyee==8.2.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
selenium-stealth==1.0.6
sqlparse==0.4.2
TikTokApi==4.0.1
urllib3==1.26.6
websockets==10.0
